Implement Saving a gig

+++ GigHub/Models/Genre.cs
@@ -4,7 +4,7 @@ namespace GigHub.Models
 {
     public class Genre
     {
-        public int Id { get; set; }
+        public byte Id { get; set; }
 
         [Required]
         [StringLength(255)]
